This commit implements the wiring of stat var date ranges from the charts to the metadata modal. It covers the following tile types: bar, highlight, map, ranking and scatter.
Shared functionality is provided in the TileUtils component. A note that while line chart (which already had this functionality applied in an earlier PR) has been updated to use the shared type, it does not use the shared utility function. This is because the line tile already gathered this date / statvar information in a way that we could make use of, a way that other tiles did not.